Fix coordinates projection issue in Map tool

+++ collect-core/src/main/java/org/openforis/collect/model/LngLat.java
@@ -2,28 +2,28 @@ package org.openforis.collect.model;
 
 public class LngLat {
 
-	private Long longitude;
-	private Long latitude;
+	private Double longitude;
+	private Double latitude;
 
-	public LngLat(Long longitude, Long latitude) {
+	public LngLat(Double longitude, Double latitude) {
 		super();
 		this.longitude = longitude;
 		this.latitude = latitude;
 	}
 
-	public Long getLongitude() {
+	public Double getLongitude() {
 		return longitude;
 	}
 
-	public void setLongitude(Long longitude) {
+	public void setLongitude(Double longitude) {
 		this.longitude = longitude;
 	}
 
-	public Long getLatitude() {
+	public Double getLatitude() {
 		return latitude;
 	}
 
-	public void setLatitude(Long latitude) {
+	public void setLatitude(Double latitude) {
 		this.latitude = latitude;
 	}	
 }
